+++ include/apex/pnl/intra_summary.hpp
@@ -6,7 +6,7 @@ struct IntraSummary {
     int run;
     int seed;
     double final_total_pnl;
-    int64_t final_inventory;
+    double final_inventory;
     double sharpe_ratio;
     double max_drawdown;
 };
